pdfprocess/requirements.txt
# PDF Processing Service - Requirements
# Deploy to Render

# Web Framework
fastapi==0.115.0
uvicorn==0.27.0
python-multipart==0.0.9

# Settings & Environment
pydantic-settings==2.12.0
python-dotenv==1.0.0

# Database
supabase==2.24.0

# HTTP Client (for webhook)
httpx==0.27.0

# PDF Processing
PyMuPDF==1.26.6
pymupdf4llm==0.2.5
PyPDF2==3.0.1

# Document Processing
python-docx==1.1.0
beautifulsoup4==4.12.3
Markdown==3.5.2
lxml==6.0.2

# Text Chunking
langchain-text-splitters==0.3.11

# Async File Operations
aiofiles==23.2.1
